data-pipeline/pyproject.toml
[project]
name = "data-collection"
version = "0.1.0"
description = "Ethereum data collection pipeline scaffold"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Crypto Narrative Hunter Team" }]
dependencies = [
  "psycopg[binary]>=3.1.0,<4",
  "psycopg-pool>=3.1.0,<4",
  "SQLAlchemy>=2.0.0,<3",
  "requests>=2.31.0,<3",
  "pandas>=2.2.2,<3",
  "pyarrow>=17.0.0,<18",
  "python-dotenv>=1.0.1,<2",
  "tqdm>=4.66.4,<5",
  "pydantic>=2.9.2,<3",
  "tenacity>=8.5.0,<9",
  "httpx>=0.27.0,<1",
  "pytest>=8.3.0,<9",
  "ruff>=0.6.0,<0.7",
  "coverage>=7.6.0,<8",
  "matplotlib>=3.7.0,<4",
  "seaborn>=0.12.0,<1",
  "scikit-learn>=1.5.0,<1.6",
  "hdbscan>=0.8.40,<0.9",
  "networkx>=3.0,<4",
  "pyyaml>=6.0,<7",
  "click>=8.1.0,<9",
  "aiohttp>=3.9.5,<4",
  "web3>=6.20.1,<8",
  "structlog>=24.1.0,<26",
  "asyncpg>=0.29.0,<1",
  "psutil>=5.9.0,<6",
]

[project.optional-dependencies]
future = [
  "aiohttp>=3.9.5,<4",
  "web3>=6.20.1,<7",
  "structlog>=24.1.0,<25",
  "asyncpg>=0.29.0,<1",
  "pytest-asyncio>=0.23.0,<1",
  "aioresponses>=0.7.6,<1",
  "gql>=3.5.0,<4",
]

[project.scripts]
data-collection = "data_collection.cli:main"

[tool.uv]
managed = true
